windows
me waiting for MSVC to stop making me write properly, deepls retarded
code: https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fmedia.istockphoto.com%2Fphotos%2Fskeleton-reaching-for-a-window-picture-id155154794%3Fk%3D6%26m%3D155154794%26s%3D612x612%26w%3D0%26h%3DKNIrAKdhAsI1FCoMtSLXd99ZnP9MO1zZj8VkThTvwVs%3D&f=1&nofb=1

+++ src/Common.cpp
@@ -102,7 +102,7 @@ TEST_CASE("Application::IsOutdated (version check)") {
         for (uint8_t Patch = 0; Patch < 10; ++Patch) {
             for (uint8_t Minor = 0; Minor < 10; ++Minor) {
                 for (uint8_t Major = 0; Major < 10; ++Major) {
-                    CHECK(Application::IsOutdated({ Major, Minor, Patch }, { Major, Minor, Patch + 1 }));
+                    CHECK(Application::IsOutdated({ uint8_t(Major), uint8_t(Minor), uint8_t(Patch) }, { uint8_t(Major), uint8_t(Minor), uint8_t(Patch + 1) }));
                 }
             }
         }
@@ -111,7 +111,7 @@ TEST_CASE("Application::IsOutdated (version check)") {
         for (uint8_t Patch = 0; Patch < 10; ++Patch) {
             for (uint8_t Minor = 0; Minor < 10; ++Minor) {
                 for (uint8_t Major = 0; Major < 10; ++Major) {
-                    CHECK(Application::IsOutdated({ Major, Minor, Patch }, { Major, Minor + 1, Patch }));
+                    CHECK(Application::IsOutdated({ uint8_t(Major), uint8_t(Minor), uint8_t(Patch) }, { uint8_t(Major), uint8_t(Minor + 1), uint8_t(Patch) }));
                 }
             }
         }
@@ -120,7 +120,7 @@ TEST_CASE("Application::IsOutdated (version check)") {
         for (uint8_t Patch = 0; Patch < 10; ++Patch) {
             for (uint8_t Minor = 0; Minor < 10; ++Minor) {
                 for (uint8_t Major = 0; Major < 10; ++Major) {
-                    CHECK(Application::IsOutdated({ Major, Minor, Patch }, { Major + 1, Minor, Patch }));
+                    CHECK(Application::IsOutdated({ uint8_t(Major), uint8_t(Minor), uint8_t(Patch) }, { uint8_t(Major + 1), uint8_t(Minor), uint8_t(Patch) }));
                 }
             }
         }
@@ -129,7 +129,7 @@ TEST_CASE("Application::IsOutdated (version check)") {
         for (uint8_t Patch = 0; Patch < 10; ++Patch) {
             for (uint8_t Minor = 0; Minor < 10; ++Minor) {
                 for (uint8_t Major = 0; Major < 10; ++Major) {
-                    CHECK(Application::IsOutdated({ Major, Minor, Patch }, { Major + 1, Minor + 1, Patch + 1 }));
+                    CHECK(Application::IsOutdated({ uint8_t(Major), uint8_t(Minor), uint8_t(Patch) }, { uint8_t(Major + 1), uint8_t(Minor + 1), uint8_t(Patch + 1) }));
                 }
             }
         }
